[ZEPPELIN-2813] refactoring
