+++ zeppelin-web/src/app/notebook/paragraph/paragraph.css
@@ -270,117 +270,6 @@ table.table-shortcut {
   font-size: 10px !important;
 }
 
-.revisions-comparator-modal-dialog {
-  width: 80%;
-  margin: 30px auto;
-  height: 80%;
-}
-
-.revisions-comparator-modal-dialog .modal-content {
-  height: 100%;
-}
-
-.revisions-comparator-modal-header {
-  min-height: 16.428571429px;
-  padding: 15px;
-  border-bottom: 1px solid #9cb4c5;
-  background-color: #3071a9;
-  border: 2px solid #3071a9;
-  border-top-left-radius: 4px;
-  border-top-right-radius: 4px;
-}
-
-.revisions-comparator-modal-header .close {
-  color: #cfcfcf;
-  opacity: 1;
-}
-
-.revisions-comparator-modal-title {
-  color: white;
-  margin-top: 2px;
-  margin-bottom: 2px;
-}
-
-.revisions-comparator-code-panel {
-  display: inline-block;
-  width: 50%;
-  float: left;
-  height: 300px;
-  overflow-y: auto;
-  max-height: 300px
-}
-
-.revisions-comparator-code-panel-title {
-  width: 50%;
-  float: left;
-  font-size: 14px;
-  padding: 5px;
-  background-color: #f5f5f5;
-}
-
-.revisions-comparator-bar {
-  margin: 10px auto;
-  width: 400px;
-}
-
-.revisions-comparator-status {
-  font-size: 12px;
-  padding-left: 10px;
-}
-
-#diffPanel {
-  height: calc(100% - 60px);
-}
-
-#diffPanel .panel-group {
-  height: inherit;
-  overflow: auto;
-}
-
-.revision-name-for-compare {
-  cursor: default;
-  overflow: hidden;
-  display: inline-block;
-  max-width: 100px;
-  width: 100px;
-  padding: 1px 10px;
-}
-
-.revisions-comparator-caret {
-  margin: 0 0 10px 6px;
-}
-
-.revisions-comparator-link, .revisions-comparator-link:hover,
-.revisions-comparator-link:visited, .revisions-comparator-link:focus {
-  text-decoration: none;
-  color: #000;
-}
-
-.revisions-comparator-forst-string {
-  display: block;
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  font-size: 12px;
-  color: grey;
-}
-
-.color-green {
-  color: green;
-}
-
-.color-red {
-  color: red;
-}
-
-.color-black {
-  color: black;
-}
-
-.color-orange {
-  color: orange;
-}
-
 /*
   Paragraph Title
 */
