--- zeppelin-web/src/app/notebook/revisions-comparator.html
@@ -1,126 +1,0 @@
-<!--
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
--->
-
-<div class="modal fade" id="revisionsComparatorModal" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true"
-     modalvisible previsiblecallback="preVisibleRevisionsComparator">
-  <div class="revisions-comparator-modal-dialog">
-    <div class="modal-content">
-      <div class="revisions-comparator-modal-header">
-        <button type="button" class="close" data-dismiss="modal"><span aria-hidden="true">&times;</span><span class="sr-only">Close</span></button>
-        <h4 class="revisions-comparator-modal-title" id="myModalLabel">Revisions comparator</h4>
-      </div>
-      <div class="revisions-comparator-bar">
-        <div class="btn-group">
-          <button type="button" ng-if="noteRevisions.length > 0"
-                  class="btn btn-sm btn-default dropdown-toggle"
-                  data-toggle="dropdown" id="firstRevisionDropdown" title="{{currentFirstRevisionForCompare}}">
-            <div class="revision-name-for-compare">{{currentFirstRevisionForCompare}}</div>
-            <span class="caret revisions-comparator-caret"></span>
-          </button>
-          <ul class="dropdown-menu pull-right" aria-labelledby="firstRevisionDropdown">
-            <li></li>
-            <li ng-repeat="revision in noteRevisions | orderBy:'time':true" class="revision">
-              <a style="cursor:pointer" ng-click="getNoteRevisionForReview(revision, 'first')">
-                <span style="display: block;">
-                  <strong>{{revision.message}}</strong>
-                </span>
-                <span class="revisionDate">
-                  <em>{{formatRevisionDate(revision.time)}}</em>
-                </span>
-              </a>
-            </li>
-          </ul>
-        </div>
-        <span>compare with</span>
-        <div class="btn-group">
-          <button type="button" ng-if="noteRevisions.length > 0"
-                  class="btn btn-sm btn-default dropdown-toggle"
-                  ng-disabled="firstNoteRevisionForCompare === null"
-                  data-toggle="dropdown" id="secondRevisionDropdown" title="{{currentSecondRevisionForCompare}}">
-            <div class="revision-name-for-compare">{{currentSecondRevisionForCompare}}</div>
-            <span class="caret revisions-comparator-caret"></span>
-          </button>
-          <ul class="dropdown-menu pull-right" aria-labelledby="secondRevisionDropdown">
-            <li ng-repeat="revision in noteRevisions | orderBy:'time':true" class="revision">
-              <a style="cursor:pointer" ng-click="getNoteRevisionForReview(revision, 'second')">
-                <span style="display: block;">
-                  <strong>{{revision.message}}</strong>
-                </span>
-                <span class="revisionDate">
-                  <em>{{formatRevisionDate(revision.time)}}</em>
-                </span>
-              </a>
-            </li>
-          </ul>
-        </div>
-      </div>
-      <div id="diffPanel">
-        <div class="panel-group">
-          <div class="panel" data-ng-repeat="p in (mergeNoteRevisionsForCompare ? mergeNoteRevisionsForCompare.added : [])">
-            <div class="panel-heading">
-              <a class="revisions-comparator-link" data-toggle="collapse" data-parent="#diffPanel" href="#{{p.paragraph.id}}">
-                <h4 class="panel-title">
-                  {{p.paragraph.id}}<strong style="padding: 5px;" ng-if="p.paragraph.title">({{p.paragraph.title}})</strong>
-                  <i class="revisions-comparator-status color-green">added</i>
-                  <i class="revisions-comparator-forst-string">{{p.firstString}}</i>
-                </h4>
-              </a>
-            </div>
-            <div id="{{p.paragraph.id}}" class="panel-collapse collapse">
-              <span class="revisions-comparator-code-panel-title">Revision: <strong>{{currentFirstRevisionForCompare}}</strong></span>
-              <span class="revisions-comparator-code-panel-title">Revision: <strong>{{currentSecondRevisionForCompare}}</strong></span>
-              <pre class="revisions-comparator-code-panel"></pre>
-              <pre class="revisions-comparator-code-panel color-green">{{p.paragraph.text}}</pre>
-            </div>
-          </div>
-          <div class="panel" data-ng-repeat="p in (mergeNoteRevisionsForCompare ? mergeNoteRevisionsForCompare.deleted : [])">
-            <div class="panel-heading">
-              <a class="revisions-comparator-link" data-toggle="collapse" data-parent="#diffPanel" href="#{{p.paragraph.id}}">
-                <h4 class="panel-title">
-                  {{p.paragraph.id}} <strong style="padding: 5px;" ng-if="p.paragraph.title">({{p.paragraph.title}})</strong>
-                  <i class="revisions-comparator-status color-red">deleted</i>
-                  <i class="revisions-comparator-forst-string">{{p.firstString}}</i>
-                </h4>
-              </a>
-            </div>
-            <div id="{{p.paragraph.id}}" class="panel-collapse collapse">
-              <span class="revisions-comparator-code-panel-title">Revision: <strong>{{currentFirstRevisionForCompare}}</strong></span>
-              <span class="revisions-comparator-code-panel-title">Revision: <strong>{{currentSecondRevisionForCompare}}</strong></span>
-              <pre class="revisions-comparator-code-panel color-red">{{p.paragraph.text}}</pre>
-              <pre class="revisions-comparator-code-panel"></pre>
-            </div>
-          </div>
-          <div class="panel" data-ng-repeat="p in (mergeNoteRevisionsForCompare ? mergeNoteRevisionsForCompare.compared : [])">
-            <div class="panel-heading">
-              <a class="revisions-comparator-link" data-toggle="collapse" data-parent="#diffPanel" href="#{{p.paragraph.id}}">
-                <h4 class="panel-title">
-                  {{p.paragraph.id}} <strong style="padding: 5px;" ng-if="p.paragraph.title">({{p.paragraph.title}})</strong>
-                  <i class="revisions-comparator-status" ng-show="p.identical">contents are identical</i>
-                  <i class="revisions-comparator-status color-orange" ng-show="!(p.identical)">there are differences</i>
-                  <i class="revisions-comparator-forst-string">{{p.firstString}}</i>
-                </h4>
-              </a>
-            </div>
-            <div id="{{p.paragraph.id}}" class="panel-collapse collapse">
-              <span class="revisions-comparator-code-panel-title">Revision: <strong>{{currentFirstRevisionForCompare}}</strong></span>
-              <span class="revisions-comparator-code-panel-title">Diff with revision: <strong>{{currentSecondRevisionForCompare}}</strong></span>
-              <pre class="revisions-comparator-code-panel">{{p.paragraph.text}}</pre>
-              <pre class="revisions-comparator-code-panel" ng-bind-html="p.diff"></pre>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
