+++ zeppelin-web/src/app/notebook/revisions-comparator/revisions-comparator.html
@@ -1,0 +1,87 @@
+<!--
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+-->
+
+<div class="revisions-comparator-bar">
+  <div class="btn-group">
+    <button type="button" ng-if="$ctrl.noteRevisions.length > 0"
+            class="btn btn-default revisions-comparator-dropdown dropdown-toggle"
+            data-toggle="dropdown" id="firstRevisionDropdown" title="{{currentFirstRevisionForCompare}}">
+      <div class="revision-name-for-compare">{{currentFirstRevisionForCompare}}</div>
+      <span class="caret revisions-comparator-caret"></span>
+    </button>
+    <ul class="dropdown-menu dropdown-menu-left" aria-labelledby="firstRevisionDropdown">
+      <li></li>
+      <li ng-repeat="revision in $ctrl.noteRevisions | orderBy:'time':true" class="revision">
+        <a style="cursor:pointer" ng-click="getNoteRevisionForReview(revision, 'first')">
+          <span style="display: block;">
+            <strong>{{revision.message}}</strong>
+          </span>
+          <span class="revisionDate">
+            <em>{{formatRevisionDate(revision.time)}}</em>
+          </span>
+        </a>
+      </li>
+    </ul>
+  </div>
+  <span>compare with</span>
+  <div class="btn-group">
+    <button type="button" ng-if="$ctrl.noteRevisions.length > 0"
+            class="btn btn-default revisions-comparator-dropdown dropdown-toggle"
+            ng-disabled="firstNoteRevisionForCompare === null"
+            data-toggle="dropdown" id="secondRevisionDropdown" title="{{currentSecondRevisionForCompare}}">
+      <div class="revision-name-for-compare">{{currentSecondRevisionForCompare}}</div>
+      <span class="caret revisions-comparator-caret"></span>
+    </button>
+    <ul class="dropdown-menu dropdown-menu-left" aria-labelledby="secondRevisionDropdown">
+      <li ng-repeat="revision in $ctrl.noteRevisions | orderBy:'time':true" class="revision">
+        <a style="cursor:pointer" ng-click="getNoteRevisionForReview(revision, 'second')">
+          <span style="display: block;">
+            <strong>{{revision.message}}</strong>
+          </span>
+          <span class="revisionDate">
+            <em>{{formatRevisionDate(revision.time)}}</em>
+          </span>
+        </a>
+      </li>
+    </ul>
+  </div>
+</div>
+<div id="diffPanel">
+  <div class="panel-group">
+    <div class="panel" data-ng-repeat="p in mergeNoteRevisionsForCompare">
+      <div class="revisions-comparator-panel-heading">
+        <a class="revisions-comparator-link" data-toggle="collapse" data-parent="#diffPanel" href="#{{p.paragraph.id}}">
+          <h4 class="panel-title">
+            {{p.paragraph.id}}<strong style="padding: 5px;" ng-if="p.paragraph.title">({{p.paragraph.title}})</strong>
+            <i ng-if="p.type === 'added'" class="revisions-comparator-status color-green">added</i>
+            <i ng-if="p.type === 'deleted'" class="revisions-comparator-status color-red">deleted</i>
+            <i ng-if="p.type === 'compared' && !(p.identical)" class="revisions-comparator-status color-orange">there are differences</i>
+            <i ng-if="p.type === 'compared' && p.identical" class="revisions-comparator-status">contents are identical</i>
+            <i class="revisions-comparator-first-string">{{p.firstString}}</i>
+          </h4>
+        </a>
+      </div>
+      <div id="{{p.paragraph.id}}" class="panel-collapse collapse">
+        <span class="revisions-comparator-code-panel-title">Revision: <strong>{{currentFirstRevisionForCompare}}</strong></span>
+        <span class="revisions-comparator-code-panel-title">Revision: <strong>{{currentSecondRevisionForCompare}}</strong></span>
+        <pre ng-if="p.type === 'added'" class="revisions-comparator-code-panel"></pre>
+        <pre ng-if="p.type === 'added'" class="revisions-comparator-code-panel color-green">{{p.paragraph.text}}</pre>
+        <pre ng-if="p.type === 'deleted'" class="revisions-comparator-code-panel color-red">{{p.paragraph.text}}</pre>
+        <pre ng-if="p.type === 'deleted'" class="revisions-comparator-code-panel"></pre>
+        <pre ng-if="p.type === 'compared'" class="revisions-comparator-code-panel">{{p.paragraph.text}}</pre>
+        <pre ng-if="p.type === 'compared'" class="revisions-comparator-code-panel" ng-bind-html="p.diff"></pre>
+      </div>
+    </div>
+  </div>
+</div>
