+++ zeppelin-web/src/index.js
@@ -42,6 +42,7 @@ import './app/interpreter/interpreter-item.directive.js'
 import './app/interpreter/widget/number-widget.directive.js'
 import './app/credential/credential.controller.js'
 import './app/configuration/configuration.controller.js'
+import './app/notebook/revisions-comparator/revisions-comparator.component.js'
 import './app/notebook/paragraph/paragraph.controller.js'
 import './app/notebook/paragraph/clipboard.controller.js'
 import './app/notebook/paragraph/resizable.directive.js'
